pyproject.toml
[project]
name = "better-job-search"
version = "2.0.0"
description = "Job market analysis with NLP pipeline, UMAP/HDBSCAN clustering, multi-agent system, and React UI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10,<3.13"
authors = [
    {name = "Sina Karaoglu"}
]
keywords = ["jobs", "nlp", "rag", "search", "clustering", "agents", "umap", "fastapi"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Text Processing :: Indexing",
]

dependencies = [
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "sentence-transformers>=2.2",
    "faiss-cpu>=1.7",
    "rank-bm25>=0.2",
    "beautifulsoup4>=4.12",
    "tqdm>=4.66",
    "pyyaml>=6.0",
    "pyarrow>=15.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
]

[project.optional-dependencies]
nlp = [
    "spacy>=3.7",
    "cymem<2.0.13",
    "keybert>=0.8",
    "rapidfuzz>=3.0",
]
clustering = [
    "umap-learn>=0.5.7",
    "hdbscan>=0.8.40",
    "numba>=0.59",
]
agents = [
    "openai>=1.0",
    "tiktoken>=0.5",
]
api = [
    "fastapi>=0.110",
    "uvicorn>=0.27",
    "websockets>=12.0",
]
llm = [
    "openai>=1.0",
]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
    "ipykernel>=6.20",
    "jupyterlab>=4.0",
]
all = [
    "better-job-search[nlp,clustering,agents,api,llm,dev]",
]

[project.scripts]
better-job-search = "src.pipeline:main"

[project.urls]
Homepage = "https://github.com/karaoglusina/better-job-search"
Repository = "https://github.com/karaoglusina/better-job-search"
Issues = "https://github.com/karaoglusina/better-job-search/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
